Action: toggle_maximize_horizontally
The built in maximze function forces the window to be fully visible so if we're
to use these we have to do some stuff when focus changes to other window.

One option is to simply unmaximize on focus-lost.

Or find out if it's possible to disable those constraints. (probably not, see
constraints in mutter)