Reduce the chance of windows passing through each other in the z-axis
When going to a scaled window set a tiny delay so the window to the left/right
have time to move out of the way before we `raise` and start the animation. The
animation time is reduced by the delay, so the animations should still finish at
the same time.

This isn't completely robust because it depends on the speed of window to the
left/right moves, which is determine by the width of the window we're ensuring.