Use somewhat proper delay calculation
This should in principle start the animation when the obscuring window have
moved margin - margin_lr frames, which is in the ballbark of what we want.

Had to divide the whole thing by two, not sure why that should've been necessary.

It might be better to use easeOutQuad though, since the other windows are
already moving.