Bugfix: PreviewedWindowNavigator ensured wrong window sometimes
To reproduce:
- activate the popup
- move the pointer over an entry
- use the keyboard to select another entry

On modifier release the keyboard-selected window get focus, but the mouse/pointer
hovered window is ensured.

Not sure why item-enter (hover) is triggered since the pointer already has
entered that item, but this is a workaround. Even if the pointer was required to
move the situation could probably be triggered by actually moving the pointer
while releasing the modifier key.