stylesheet.css
And a placeholder style for selected window in minimap.

The extension system load this file automatically. Simply set the class on a st
widget to use:

    new St.Bin({ style_class: "paper-mm-selected-window"});

The border is part of the width though (box-sizing not supported I think) so
if we're going to use borders directly around the clone some adjustment will be
needed to keep 100% aligned with the actual space.