WIP: Two working alternatives for a wrapper that let us use "frame coordinates"
Note: the allocation method is called ALOT

WindowCloneLayout based on the one in workspace.js, but the original
calulates the offset based on a static frame_rect. Unless that's updated the
clone's position will follow the source. The containers position (using the
layout manager) will _not_ follow it though. So the clone is not properly
contained inside the container..