Fix massive st_widget_get_theme_node warnings from minimap/multimap
   st_widget_get_theme_node called on the widget [0x3f47ec0 StWidget "window-clone-container"] which is not in the stage.

This warning is emitted if any information dependent on a widget's style is
requested before the widget is attached to the stage.

In this case the culprit was that St.Widget.get_preferred_width/height call
st_widget_get_theme_node even when its layoutmanager dictate the preferred size.

It's convenient to be able to use the size of the window clones prior to
attachment, and the size should not depend on any styling.

We don't use any St functionality in the clone wrapper so simply use a regular
Clutter actor.