Support most of our own actions in "supermode"
As a consequence the metawindows now keep track of their destination
when animated by `move_to`.

Ideally the minimap/multimap should listen for changes to the space
and react accordingly instead of doing a manual sync like this.

"most actions" means actions that is registered as PER_WINDOW and
doesn't mutate the space.