New paged windows should never be `on top`
Windows on the stack should never be on top, windows not on the stack will never
overlap. Which means that alaways on top is never wanted for ordered windows.

The issue is noticed when attaching a scratch window to the page strip, as it
will continue to be always on top, making the stack overlap visible windows.