dual monitor: Enable moving windows to shared monitors
When there's only workspaces on the primary display, moving a window to another
monitor makes it `stick`. In the past this would make it part of the scratch
layer, but that's not the case anymore, so it would just be detached and
attached to the workspace again.

This makes it possible to simply drag a window from a space into another monitor
and it will detach and float on the monitor.