performance: Minimize lower/raise calls
The bottleneck when calling ensure seems to be calling lower on every window on
the left/right stack.

This is a quick fix, where we simply call raise on the top of the left/right
stack, and the visible windows.

A consequence of this change is the left/right stacks might be above a visible
window for a short time, but it's not very noticeable.