Move left/right stack logic to ensure and propagate functions
When we propagate the position forward/backward we check the desired position
against the left/right limits and stack the rest of the windows if it's the
case.

This also fixes the stack performance issue in a cleaner way. We simply raise
all non-stacked windows while propagating, instead of doing a fix-up afterwards.