Focus the right neighbour when closing a window
Mutter hands focus to the "highest" window when a window is closed. This means
we need to raise windows from the edges in towards the selected window, making
the neighbours of the selected window the next highest windows.

In particular since we propagate backwards last, the right neighbour will get
priority.