Don't execute toggle-scratch-layer when previewing windows
It doesn't make much sense to show the scratch layer when previewing. In
addition if using `super+escape` as the binding this makes it impossible to
abort the preview.

In the long run it's probably better to make a whitelist of bindings that's
appropriate to execute while a preview is active.