Use clones for all animation
Do all animations with clones, hiding window actors. When animations are
complete we toggle back to window actors if the window isn't stacked.

We put the clones in their own Space specific container in Main.uiGroup. Above
the BackgroundGroup (moved to the bottom of the uiGroup) and below the
window_group.

Also remove the then obsolete stackWindow.