Fix broken transient windows:
Some dialog windows doesn't set the transient hint. Without that hint it's hard
to sanely treat the window as eg. a dialog so we simply add it. (eg. gnome-shell
extension preference windows)

In the clone-stack branch this becomes more apparent that before I think.

We could handle this case-by-case in kludges.js instead?

We could try to guess the parent window and set transient_of?
But not sure how hard that is both technically (can we set the hint?) and
logically (can we reliable guess the correct window?) I think notion does
something like that.

We could simply set "above" too?