Contain all spaces in their own clipping container
We create an actor spanning the whole screen, `Spaces.spaceContainer` which
holds all spaces. A `Space` is now first contained in a `Space.clip` container
which clips away anything that protrudes from the monitor.

The main advantage over having a per monitor clipping actor is getting
flexibility in moving a space around the screen without changing it's monitor.

We also hook up the `monitors-changed` signal, putting the active workpace on
the primary display, and populating the other monitors with the rest of the
workspace mru.

This also prevents a bug where we could end up moving a BackgroundActor.