Let the `Space::move-done` signal take care of showing WindowActors
We store windows that should be visible, ie. their WindowActor should be shown,
in a `Space.visible` array. When done with animation we emit `move-done` and the
space will take care of replacing the clones with WindowActors.

Navigator integration is a bit awkward as we simply rely on a `Space.delay` to
indicate if we're in a non-moving state and can commit the WindowActors
immediately when finishing a navigation.