Fix dynamic workspaces again
We need to update the most recently used workspace stack when adding and
removing workspaces.

To make this less of a pain we make spaces.stack track all spaces, not just the
ones that aren't visible. Prefering instead to filter out the visible spaces on
`selectSpace`, this is a lot cleaner.