Avoid getting stuck in preview state after swiping
`space.moving` is obsolete so remove it. We already guard against starting new
animations with the same target in `move_to`, which means keeping track of the
«moving» window is unnecessary.

I also cleaned up the code a bit, removed a bunch of unused parameters.

The bug is a good indication why `space.moving` was pretty ad-hoc and bad:

Returning when `space.moving == metaWindow` assumed there was already a tween
running, which we didn't want to replace. Swiping breaks this assumption by
removing any existing tweens. So a swiping to a already «moving» window would
result in ensureViewport returning early, never calling moveDone.