Space.layout: only tween if needed
When clones are active windows can't take pointer input. 9389d90cc made it
possible to only activate windows who's clone isn't animating.

We can take advantage of this by not forcing `ensureViewport` when a grab ends
triggering a moveDone activating the grabbed window immediately if it ended in
an already correct position.