toggle-maximize-width: more robust detection of "is maximized"
Simply check if the window is currently wide enough to be considered maximized.

Before we used `unmaximizedRect` to track maximized status. That is of course
brittle since the window can change size in many way. We cleared
`unmaximizedRect` in cycle-width (but it has been buggy for along time :P)

Note: using the size-changed signal is not trivial for this purpose: the signal
doesn't provide the previous size and it is async in wayland and synchronous in
X11.. (after running move_resize_frame)