Don't use `utils` in `extension.js`
If utils fails to load we're not getting proper error messages as we're using it
outside a try block. Use `log` instead.

SyntaxErrors still doesn't return a proper trace unfortunately, though that's
perhaps better fixed in tooling. The alternative would be to run `Reflect.parse`
on the module file in the catch block returning that error instead...