Regression: Prev commit (af3623e) broke new window animation
Try removing the Navigator.navigating and inPreview checks. In those cases the
clone should be visible.

Without those checks (together with the clone.get_parent check?) dialogs created
during animation is not hidden.

Also note that the unmapped description in af362e was wrong - unmapped is only
true a short period after the window is created.