Use the clone's stack position instead of windows'
We need to raise one of the focused window neighbours preemptively to get sane
focus after closing a window (ie. one of the neighbours should always get
focus).

However we also use the windows' stacking order to decide eg. which window in
vertical tiling to select. Raising the neighbours on focus can easily break this
behavior.

If we use the clones stacking order instead we avoid these issues.