liveAltTab: run moveDone directly
Needed when the `to` window is in the scratch. In that case focus_handler
doesn't cause a moveDone.

The most obvious effect (X11) of not calling moveDone is that the background is
still tracked and the scratch window is thus unresponsive (ref: #80). It
also makes the tiled windows unresponsive to the first click since the window
actors will be hidden (wayland too I think) - the tiling will pick up the click
and activate the clicked window though.