Fix weird minimize animation of scratch windows with remembered position
The scratch window would first be moved to the remembered position - then
animated "down".

Also ensures that windows minimized when navigating get a proper position.
(restoring a window that was minimized when it was non-placeable results in the
window being out-of-view)

Note: Use 'effect-completed' to detect when a the animation is done - then
resize/move. Could also have stored the target position, but that complicates
all the places we want to preview a minimized window.