Use a shadowed border to frame spaces in the carousel
A colored border is added to frame the space when using the workspace carousel.

This helps to separate the workspaces visually. Especially when using a
shared image background for workspaces and the gnome shell background.

We also make the background the same size as the monitor making the overview
transition smooth when using a single background.

During regular use the border is hidden, to avoid shadows at the edges of the
monitor.

closes #86