insertWindow: Do not handle tiled windows that are stuck
More only-on-primary work.

When a secondary monitor is to the left/right of primary, tiled windows become
`stuck` when they're pushed into the secondary monitor. Eg. if `A` and `B` is
tiled `B` will become `stuck` by entiring the secondary monitor:

|-----primary-----|----secondary------|
| |-------------| | |-------------|   |
| |             | | |             |   |
| |             | | |             |   |
| |      A      | | |      B      |   |
| |             | | |             |   |
| |_____________| | |_____________|   |
|_________________|___________________|

Handle this edge case by not doing anything with such windows.