performance: Deactivate clones when hidden
Having an active clone can cause some overhead, even when the window is
hidden. This can be observed if eg. playing a video in a hidden
workspace: the gnome-shell process will use around 10% cpu (at least on
my machine).

While we don't want to destroy and create clones on demand we simply set
null out the clone's source, which have the desired effect. This seems
to work smoothly.