Add instruction on windows matching
I was trying to find a way to put a window that has no defined `wm_class` to scratch layer. I had thought of using `title` to match the window and only realised that it had already been implemented without being mentioned in the README.

This PR adds some hints to such an approach for window matching.