Warp global pointer to follow touch event
This is an incomplete solution, as the drag-and-drop event isn't
properly finished in touch cases, requiring a trackpad swipe to "unlock"
the situation.

Also, while simple, moving the global pointer goes against expected
touch behavior in GNOME, which is that touch events do not move the
mouse. This could be adressed by memorizing and restoring the pointer as
the event starts.