More straightforward direction mappings
We have two direction enums, which basically mean the same thing.
Previously we were figuring out which monitor we need to go to depending
on which way we were going out of bounds, but that's unnecessary, since
we already know which direction we're heading.

If this mapping is needed elsewhere I'll move it into a util or
something.